Log only unexpected exceptions in ExceptionHandlerMiddleware
UnhandledExceptionLoggingBehaviour (Application assembly): catches the unhandled exception, occured in the MediatR pipeline, logs this exception, and then simply throws it.

ExceptionHandlerMiddleware (Presentation.API assembly): catches all exceptions, occured in the ASP.NET Core pipeline, and then simply returns a response to user (with an appropriate HTTP status-code). In this middleware we handle some concrete exception types, but also there is a special handler for unexpected exception (i.e., exception of unexpected type). This exception should be logged just in case.