Remove post only if it has no comments (validation)
When user removes a post, we check for comments of this post. Post can be removed only if it has no comments.
This was made using FluentValidation.

Before this commit we had this check inside of DeletePostCommand, but it's better to make it using FluentValidation - place all validation rules in one place according to the following answer: https://stackoverflow.com/a/51562385/11285108