Add column UserId to Post, PostRate, Comment and CommentRate entities
We have to separate databases - Primary and Identity.
Entities 'Post', 'PostRate', 'Comment' and 'CommentRate' are related to the Primary-database. We added a column UserId to each of them, which contains the id of a user, stored in the Identity-database.

Also we disabled the seeder, because it seeds posts, without setting 'UserId' property, which will throw an exception.