Make IdentityUserService non-generic, add TRole to IIdentityUserService
IIdentityUserService is an interface that provides APIs to interact with user storage and user roles. By now it allows to interact only with user storage, but in the future it will allow to work with user roles. So, we added a generic-type TRole to this interface.

IdentityUserService is an implementor of IIdentityUserService. Inside of this type we need to work with ApplicationUser and ApplicationRole directly. Because of this we made it non-generic in this commit.