Remove CommentRateId from CommentRateDto, PostRateId from PostRateDto
When someone queries rates of a comment, we return a collection of elements of type CommentRateDto. We removed the property CommentRateId because there is no need for this property. To remove a comment-rate we need to know the comment-id and the user-id, but not the comment-rate-id.
The same thing with PostRateDto and its property PostRateId.